+++ src/Mathematics.NET/DifferentialGeometry/TensorFieldF3.cs
@@ -1,4 +1,4 @@
-// <copyright file="FMMetricTensorField2x2.cs" company="Mathematics.NET">
+// <copyright file="TensorFieldF3.cs" company="Mathematics.NET">
 // Mathematics.NET
 // https://github.com/HamletTanyavong/Mathematics.NET
 //
@@ -26,54 +26,63 @@
 // </copyright>
 
 using System.Numerics;
+using System.Runtime.CompilerServices;
 using Mathematics.NET.AutoDiff;
 using Mathematics.NET.DifferentialGeometry.Abstractions;
 using Mathematics.NET.LinearAlgebra;
+using static Mathematics.NET.DifferentialGeometry.Buffers;
 
 namespace Mathematics.NET.DifferentialGeometry;
 
-/// <summary>Represents a 2x2 metric tensor field.</summary>
+/// <summary>Represents a rank-one tensor field with three elements.</summary>
 /// <typeparam name="TDN">A type that implements <see cref="IDual{TDN, TN, U, V}"/>.</typeparam>
 /// <typeparam name="TN">A type that implements <see cref="IComplex{T, U, V}"/> and <see cref="IDifferentiableFunctions{T}"/>.</typeparam>
 /// <typeparam name="U">A type that implements <see cref="IBinaryFloatingPointIeee754{TSelf}"/> and <see cref="IMinMaxValue{TSelf}"/>.</typeparam>
+/// <typeparam name="TIP">The position of the index of the tensor.</typeparam>
 /// <typeparam name="TPI">The index of the point on the manifold.</typeparam>
-public class FMMetricTensorField2x2<TDN, TN, U, TPI> : FMTensorField2x2<TDN, TN, U, Lower, Lower, TPI>
+public class TensorFieldF3<TDN, TN, U, TIP, TPI> : TensorField<TN, U, TPI>
     where TDN : IDual<TDN, TN, U, U>
     where TN : IComplex<TN, U, U>, IDifferentiableFunctions<TN>
     where U : IBinaryFloatingPointIeee754<U>, IMinMaxValue<U>
+    where TIP : IIndexPosition
     where TPI : IIndex
 {
-    /// <summary>Compute the value of the metric tensor at a specific point on the manifold.</summary>
-    /// <typeparam name="TI1N">The name of the first index.</typeparam>
-    /// <typeparam name="TI2N">The name of the second index.</typeparam>
-    /// <param name="point">A point on the manifold.</param>
-    /// <returns>A metric tensor.</returns>
-    public new MetricTensor<Matrix2x2<TN, U>, TN, U, Lower, TI1N, TI2N> Compute<TI1N, TI2N>(AutoDiffTensor2<TDN, TN, U, TPI> point)
-        where TI1N : IIndexName
-        where TI2N : IIndexName
+    private protected FMTensor3Buffer3<TDN, TN, U, TPI> _buffer;
+
+    public TensorFieldF3() { }
+
+    public Func<AutoDiffTensor3<TDN, TN, U, TPI>, TDN> this[int i]
     {
-        Matrix2x2<TN, U> result = new();
-        for (int i = 0; i < 2; i++)
-        {
-            for (int j = 0; j < 2; j++)
-            {
-                if (_buffer[i][j] is Func<AutoDiffTensor2<TDN, TN, U, TPI>, TDN> function)
-                    result[i, j] = function(point).D0;
-            }
-        }
-        return new MetricTensor<Matrix2x2<TN, U>, TN, U, Lower, TI1N, TI2N>(result);
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        get => _buffer[i];
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        set => _buffer[i] = value;
     }
 
-    /// <summary>Compute the value of the inverse metric tensor at a specific point on the manifold.</summary>
-    /// <typeparam name="TI1N">The name of the first index.</typeparam>
-    /// <typeparam name="TI2N">The name of the second index.</typeparam>
-    /// <param name="point">A point on the manifold.</param>
-    /// <returns>An inverse metric tensor.</returns>
-    public MetricTensor<Matrix2x2<TN, U>, TN, U, Upper, TI1N, TI2N> ComputeInverse<TI1N, TI2N>(AutoDiffTensor2<TDN, TN, U, TPI> point)
-        where TI1N : IIndexName
-        where TI2N : IIndexName
+    /// <inheritdoc cref="TensorFieldF2{TDN, TN, U, TIP, TPI}.Compute{TIN}(AutoDiffTensor2{TDN, TN, U, TPI})"/>
+    public Tensor<Vector3<TN, U>, TN, U, Index<TIP, TIN>> Compute<TIN>(AutoDiffTensor3<TDN, TN, U, TPI> point)
+        where TIN : IIndexName
     {
-        var value = Compute<TI1N, TI2N>(point);
-        return value.Inverse<TI1N, TI2N>();
+        Vector3<TN, U> result = new();
+        for (int i = 0; i < 3; i++)
+        {
+            if (_buffer[i] is Func<AutoDiffTensor3<TDN, TN, U, TPI>, TDN> function)
+                result[i] = function(point).D0;
+        }
+        return new Tensor<Vector3<TN, U>, TN, U, Index<TIP, TIN>>(result);
+    }
+}
+
+internal static partial class Buffers
+{
+    [InlineArray(3)]
+    internal struct FMTensor3Buffer3<TDN, TN, U, TPI>
+        where TDN : IDual<TDN, TN, U, U>
+        where TN : IComplex<TN, U, U>, IDifferentiableFunctions<TN>
+        where U : IBinaryFloatingPointIeee754<U>, IMinMaxValue<U>
+        where TPI : IIndex
+    {
+        private Func<AutoDiffTensor3<TDN, TN, U, TPI>, TDN> _element;
     }
 }
