+++ src/Mathematics.NET/DifferentialGeometry/TensorFieldF4x4.cs
@@ -1,4 +1,4 @@
-// <copyright file="FMTensorField2x2.cs" company="Mathematics.NET">
+// <copyright file="TensorFieldF4x4.cs" company="Mathematics.NET">
 // Mathematics.NET
 // https://github.com/HamletTanyavong/Mathematics.NET
 //
@@ -34,14 +34,14 @@ using static Mathematics.NET.DifferentialGeometry.Buffers;
 
 namespace Mathematics.NET.DifferentialGeometry;
 
-/// <summary>Represents a rank-two tensor field with four elements.</summary>
+/// <summary>Represents a rank-two tensor with 16 elements.</summary>
 /// <typeparam name="TDN">A type that implements <see cref="IDual{TDN, TN, U, V}"/>.</typeparam>
 /// <typeparam name="TN">A type that implements <see cref="IComplex{T, U, V}"/> and <see cref="IDifferentiableFunctions{T}"/>.</typeparam>
 /// <typeparam name="U">A type that implements <see cref="IBinaryFloatingPointIeee754{TSelf}"/> and <see cref="IMinMaxValue{TSelf}"/>.</typeparam>
 /// <typeparam name="TI1P">The position of the first index of the tensor.</typeparam>
 /// <typeparam name="TI2P">The position of the second index of the tensor.</typeparam>
 /// <typeparam name="TPI">The index of the point on the manifold.</typeparam>
-public class FMTensorField2x2<TDN, TN, U, TI1P, TI2P, TPI> : TensorField<TN, U, TPI>
+public class TensorFieldF4x4<TDN, TN, U, TI1P, TI2P, TPI> : TensorField<TN, U, TPI>
     where TDN : IDual<TDN, TN, U, U>
     where TN : IComplex<TN, U, U>, IDifferentiableFunctions<TN>
     where U : IBinaryFloatingPointIeee754<U>, IMinMaxValue<U>
@@ -49,11 +49,11 @@ public class FMTensorField2x2<TDN, TN, U, TI1P, TI2P, TPI> : TensorField<TN, U, 
     where TI2P : IIndexPosition
     where TPI : IIndex
 {
-    private protected FMTensor2Buffer2x2<TDN, TN, U, TPI> _buffer;
+    private protected FMTensor4Buffer4x4<TDN, TN, U, TPI> _buffer;
 
-    public FMTensorField2x2() { }
+    public TensorFieldF4x4() { }
 
-    public Func<AutoDiffTensor2<TDN, TN, U, TPI>, TDN> this[int i, int j]
+    public Func<AutoDiffTensor4<TDN, TN, U, TPI>, TDN> this[int i, int j]
     {
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         get => _buffer[i][j];
@@ -62,37 +62,33 @@ public class FMTensorField2x2<TDN, TN, U, TI1P, TI2P, TPI> : TensorField<TN, U, 
         set => _buffer[i][j] = value;
     }
 
-    /// <summary>Compute the value of the tensor at a specified point.</summary>
-    /// <typeparam name="TI1N">The name of the first index.</typeparam>
-    /// <typeparam name="TI2N">The name of the second index.</typeparam>
-    /// <param name="point">A point.</param>
-    /// <returns>The value of the tensor at the specified point.</returns>
-    public Tensor<Matrix2x2<TN, U>, TN, U, Index<TI1P, TI1N>, Index<TI2P, TI2N>> Compute<TI1N, TI2N>(AutoDiffTensor2<TDN, TN, U, TPI> point)
+    /// <inheritdoc cref="TensorFieldF2x2{TDN, TN, U, TI1P, TI2P, TPI}.Compute{TI1N, TI2N}(AutoDiffTensor2{TDN, TN, U, TPI})"/>
+    public Tensor<Matrix4x4<TN, U>, TN, U, Index<TI1P, TI1N>, Index<TI2P, TI2N>> Compute<TI1N, TI2N>(AutoDiffTensor4<TDN, TN, U, TPI> point)
         where TI1N : IIndexName
         where TI2N : IIndexName
     {
-        Matrix2x2<TN, U> result = new();
-        for (int i = 0; i < 2; i++)
+        Matrix4x4<TN, U> result = new();
+        for (int i = 0; i < 4; i++)
         {
-            for (int j = 0; j < 2; j++)
+            for (int j = 0; j < 4; j++)
             {
-                if (_buffer[i][j] is Func<AutoDiffTensor2<TDN, TN, U, TPI>, TDN> function)
+                if (_buffer[i][j] is Func<AutoDiffTensor4<TDN, TN, U, TPI>, TDN> function)
                     result[i, j] = function(point).D0;
             }
         }
-        return new Tensor<Matrix2x2<TN, U>, TN, U, Index<TI1P, TI1N>, Index<TI2P, TI2N>>(result);
+        return new Tensor<Matrix4x4<TN, U>, TN, U, Index<TI1P, TI1N>, Index<TI2P, TI2N>>(result);
     }
 }
 
 internal static partial class Buffers
 {
-    [InlineArray(2)]
-    public struct FMTensor2Buffer2x2<TDN, TN, U, TPI>
+    [InlineArray(4)]
+    public struct FMTensor4Buffer4x4<TDN, TN, U, TPI>
         where TDN : IDual<TDN, TN, U, U>
         where TN : IComplex<TN, U, U>, IDifferentiableFunctions<TN>
         where U : IBinaryFloatingPointIeee754<U>, IMinMaxValue<U>
         where TPI : IIndex
     {
-        private FMTensor2Buffer2<TDN, TN, U, TPI> _element;
+        private FMTensor4Buffer4<TDN, TN, U, TPI> _element;
     }
 }
