+++ src/Mathematics.NET/DifferentialGeometry/TensorFieldR3x3.cs
@@ -1,4 +1,4 @@
-// <copyright file="RMTensorField2x2.cs" company="Mathematics.NET">
+// <copyright file="TensorFieldR3x3.cs" company="Mathematics.NET">
 // Mathematics.NET
 // https://github.com/HamletTanyavong/Mathematics.NET
 //
@@ -34,14 +34,14 @@ using static Mathematics.NET.DifferentialGeometry.Buffers;
 
 namespace Mathematics.NET.DifferentialGeometry;
 
-/// <summary>Represents a rank-two tensor field with 4 elements.</summary>
+/// <summary>Represents a rank-two tensor field with 9 elements.</summary>
 /// <typeparam name="TT">A type that implements <see cref="ITape{T, U}"/>.</typeparam>
 /// <typeparam name="TN">A type that implements <see cref="IComplex{T, U, V}"/> and <see cref="IDifferentiableFunctions{T}"/>.</typeparam>
 /// <typeparam name="U">A type that implements <see cref="IBinaryFloatingPointIeee754{TSelf}"/> and <see cref="IMinMaxValue{TSelf}"/>.</typeparam>
 /// <typeparam name="TI1P">The position of the first index of the tensor.</typeparam>
 /// <typeparam name="TI2P">The position of the second index of the tensor.</typeparam>
 /// <typeparam name="TPI">The index of the point on the manifold.</typeparam>
-public class RMTensorField2x2<TT, TN, U, TI1P, TI2P, TPI> : TensorField<TN, U, TPI>
+public class TensorFieldR3x3<TT, TN, U, TI1P, TI2P, TPI> : TensorField<TN, U, TPI>
     where TT : ITape<TN, U>
     where TN : IComplex<TN, U, U>, IDifferentiableFunctions<TN>
     where U : IBinaryFloatingPointIeee754<U>, IMinMaxValue<U>
@@ -49,11 +49,11 @@ public class RMTensorField2x2<TT, TN, U, TI1P, TI2P, TPI> : TensorField<TN, U, T
     where TI2P : IIndexPosition
     where TPI : IIndex
 {
-    private protected RMTensor2Buffer2x2<TT, TN, U, TPI> _buffer;
+    private protected RMTensor3Buffer3x3<TT, TN, U, TPI> _buffer;
 
-    public RMTensorField2x2() { }
+    public TensorFieldR3x3() { }
 
-    public Func<TT, AutoDiffTensor2<TN, U, TPI>, Variable<TN, U>> this[int i, int j]
+    public Func<TT, AutoDiffTensor3<TN, U, TPI>, Variable<TN, U>> this[int i, int j]
     {
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         get => _buffer[i][j];
@@ -62,42 +62,37 @@ public class RMTensorField2x2<TT, TN, U, TI1P, TI2P, TPI> : TensorField<TN, U, T
         set => _buffer[i][j] = value;
     }
 
-    /// <summary>Compute the value of the tensor at a specific point on the manifold.</summary>
-    /// <typeparam name="TI1N">The name of the first index.</typeparam>
-    /// <typeparam name="TI2N">The name of the second index.</typeparam>
-    /// <param name="tape">A gradient or Hessian tape.</param>
-    /// <param name="point">A point on the manifold.</param>
-    /// <returns>A rank-two tensor.</returns>
-    public Tensor<Matrix2x2<TN, U>, TN, U, Index<TI1P, TI1N>, Index<TI2P, TI2N>> Compute<TI1N, TI2N>(TT tape, AutoDiffTensor2<TN, U, TPI> point)
+    /// <inheritdoc cref="TensorFieldR2x2{TT, TN, U, TI1P, TI2P, TPI}.Compute{TI1N, TI2N}(TT, AutoDiffTensor2{TN, U, TPI})"/>
+    public Tensor<Matrix3x3<TN, U>, TN, U, Index<TI1P, TI1N>, Index<TI2P, TI2N>> Compute<TI1N, TI2N>(TT tape, AutoDiffTensor3<TN, U, TPI> x)
         where TI1N : IIndexName
         where TI2N : IIndexName
     {
         tape.IsTracking = false;
 
-        Matrix2x2<TN, U> result = new();
-        for (int i = 0; i < 2; i++)
+        Matrix3x3<TN, U> result = new();
+        for (int i = 0; i < 3; i++)
         {
-            for (int j = 0; j < 2; j++)
+            for (int j = 0; j < 3; j++)
             {
-                if (_buffer[i][j] is Func<TT, AutoDiffTensor2<TN, U, TPI>, Variable<TN, U>> function)
-                    result[i, j] = function(tape, point).Value;
+                if (_buffer[i][j] is Func<TT, AutoDiffTensor3<TN, U, TPI>, Variable<TN, U>> function)
+                    result[i, j] = function(tape, x).Value;
             }
         }
 
         tape.IsTracking = true;
-        return new Tensor<Matrix2x2<TN, U>, TN, U, Index<TI1P, TI1N>, Index<TI2P, TI2N>>(result);
+        return new Tensor<Matrix3x3<TN, U>, TN, U, Index<TI1P, TI1N>, Index<TI2P, TI2N>>(result);
     }
 }
 
 internal static partial class Buffers
 {
-    [InlineArray(2)]
-    internal struct RMTensor2Buffer2x2<TT, TN, U, TPI>
+    [InlineArray(3)]
+    internal struct RMTensor3Buffer3x3<TT, TN, U, TPI>
         where TT : ITape<TN, U>
         where TN : IComplex<TN, U, U>, IDifferentiableFunctions<TN>
         where U : IBinaryFloatingPointIeee754<U>, IMinMaxValue<U>
         where TPI : IIndex
     {
-        private RMTensor2Buffer2<TT, TN, U, TPI> _element;
+        private RMTensor3Buffer3<TT, TN, U, TPI> _element;
     }
 }
