+++ src/Mathematics.NET/DifferentialGeometry/TensorFieldR4.cs
@@ -1,4 +1,4 @@
-// <copyright file="RMTensorField4x4.cs" company="Mathematics.NET">
+// <copyright file="TensorFieldR4.cs" company="Mathematics.NET">
 // Mathematics.NET
 // https://github.com/HamletTanyavong/Mathematics.NET
 //
@@ -34,65 +34,55 @@ using static Mathematics.NET.DifferentialGeometry.Buffers;
 
 namespace Mathematics.NET.DifferentialGeometry;
 
-/// <summary>Represents a rank-two tensor field with 16 elements.</summary>
+/// <summary>Represents a rank-one tensor field with four elements.</summary>
 /// <typeparam name="TT">A type that implements <see cref="ITape{T, U}"/>.</typeparam>
 /// <typeparam name="TN">A type that implements <see cref="IComplex{T, U, V}"/> and <see cref="IDifferentiableFunctions{T}"/>.</typeparam>
 /// <typeparam name="U">A type that implements <see cref="IBinaryFloatingPointIeee754{TSelf}"/> and <see cref="IMinMaxValue{TSelf}"/>.</typeparam>
-/// <typeparam name="TI1P">The position of the first index of the tensor.</typeparam>
-/// <typeparam name="TI2P">The position of the second index of the tensor.</typeparam>
-/// <typeparam name="TPI">The index of the point on the manifold.</typeparam>
-public class RMTensorField4x4<TT, TN, U, TI1P, TI2P, TPI> : TensorField<TN, U, TPI>
+/// <typeparam name="TIP">An index position.</typeparam>
+/// <typeparam name="TPI">An index.</typeparam>
+public class TensorFieldR4<TT, TN, U, TIP, TPI> : TensorField<TN, U, TPI>
     where TT : ITape<TN, U>
     where TN : IComplex<TN, U, U>, IDifferentiableFunctions<TN>
     where U : IBinaryFloatingPointIeee754<U>, IMinMaxValue<U>
-    where TI1P : IIndexPosition
-    where TI2P : IIndexPosition
+    where TIP : IIndexPosition
     where TPI : IIndex
 {
-    private protected RMTensor4Buffer4x4<TT, TN, U, TPI> _buffer;
+    private RMTensor4Buffer4<TT, TN, U, TPI> _buffer;
 
-    public RMTensorField4x4() { }
+    public TensorFieldR4() { }
 
-    public Func<TT, AutoDiffTensor4<TN, U, TPI>, Variable<TN, U>> this[int i, int j]
+    public Func<TT, AutoDiffTensor4<TN, U, TPI>, Variable<TN, U>> this[int i]
     {
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        get => _buffer[i][j];
+        get => _buffer[i];
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        set => _buffer[i][j] = value;
+        set => _buffer[i] = value;
     }
 
-    /// <inheritdoc cref="RMTensorField2x2{TT, TN, U, TI1P, TI2P, TPI}.Compute{TI1N, TI2N}(TT, AutoDiffTensor2{TN, U, TPI})"/>
-    public Tensor<Matrix4x4<TN, U>, TN, U, Index<TI1P, TI1N>, Index<TI2P, TI2N>> Compute<TI1N, TI2N>(TT tape, AutoDiffTensor4<TN, U, TPI> point)
-        where TI1N : IIndexName
-        where TI2N : IIndexName
+    /// <inheritdoc cref="TensorFieldR2{TT, TN, U, TIP, TPI}.Compute{TIN}(TT, AutoDiffTensor2{TN, U, TPI})"/>
+    public Tensor<Vector4<TN, U>, TN, U, Index<TIP, TIN>> Compute<TIN>(TT tape, AutoDiffTensor4<TN, U, TPI> point)
+        where TIN : IIndexName
     {
-        tape.IsTracking = false;
-
-        Matrix4x4<TN, U> result = new();
+        Vector4<TN, U> result = new();
         for (int i = 0; i < 4; i++)
         {
-            for (int j = 0; j < 4; j++)
-            {
-                if (_buffer[i][j] is Func<TT, AutoDiffTensor4<TN, U, TPI>, Variable<TN, U>> function)
-                    result[i, j] = function(tape, point).Value;
-            }
+            if (_buffer[i] is Func<TT, AutoDiffTensor4<TN, U, TPI>, Variable<TN, U>> function)
+                result[i] = function(tape, point).Value;
         }
-
-        tape.IsTracking = true;
-        return new Tensor<Matrix4x4<TN, U>, TN, U, Index<TI1P, TI1N>, Index<TI2P, TI2N>>(result);
+        return new Tensor<Vector4<TN, U>, TN, U, Index<TIP, TIN>>(result);
     }
 }
 
 internal static partial class Buffers
 {
     [InlineArray(4)]
-    internal struct RMTensor4Buffer4x4<TT, TN, U, TPI>
+    internal struct RMTensor4Buffer4<TT, TN, U, TPI>
         where TT : ITape<TN, U>
         where TN : IComplex<TN, U, U>, IDifferentiableFunctions<TN>
         where U : IBinaryFloatingPointIeee754<U>, IMinMaxValue<U>
         where TPI : IIndex
     {
-        private RMTensor4Buffer4<TT, TN, U, TPI> _element;
+        private Func<TT, AutoDiffTensor4<TN, U, TPI>, Variable<TN, U>> _element;
     }
 }
