+++ src/Mathematics.NET/DifferentialGeometry/TensorFieldR4x4.cs
@@ -1,4 +1,4 @@
-// <copyright file="RMTensorField3.cs" company="Mathematics.NET">
+// <copyright file="TensorFieldR4x4.cs" company="Mathematics.NET">
 // Mathematics.NET
 // https://github.com/HamletTanyavong/Mathematics.NET
 //
@@ -34,55 +34,65 @@ using static Mathematics.NET.DifferentialGeometry.Buffers;
 
 namespace Mathematics.NET.DifferentialGeometry;
 
-/// <summary>Represents a rank-one tensor field with three elements.</summary>
+/// <summary>Represents a rank-two tensor field with 16 elements.</summary>
 /// <typeparam name="TT">A type that implements <see cref="ITape{T, U}"/>.</typeparam>
 /// <typeparam name="TN">A type that implements <see cref="IComplex{T, U, V}"/> and <see cref="IDifferentiableFunctions{T}"/>.</typeparam>
 /// <typeparam name="U">A type that implements <see cref="IBinaryFloatingPointIeee754{TSelf}"/> and <see cref="IMinMaxValue{TSelf}"/>.</typeparam>
-/// <typeparam name="TIP">An index position.</typeparam>
-/// <typeparam name="TPI">An index.</typeparam>
-public class RMTensorField3<TT, TN, U, TIP, TPI> : TensorField<TN, U, TPI>
+/// <typeparam name="TI1P">The position of the first index of the tensor.</typeparam>
+/// <typeparam name="TI2P">The position of the second index of the tensor.</typeparam>
+/// <typeparam name="TPI">The index of the point on the manifold.</typeparam>
+public class TensorFieldR4x4<TT, TN, U, TI1P, TI2P, TPI> : TensorField<TN, U, TPI>
     where TT : ITape<TN, U>
     where TN : IComplex<TN, U, U>, IDifferentiableFunctions<TN>
     where U : IBinaryFloatingPointIeee754<U>, IMinMaxValue<U>
-    where TIP : IIndexPosition
+    where TI1P : IIndexPosition
+    where TI2P : IIndexPosition
     where TPI : IIndex
 {
-    private RMTensor3Buffer3<TT, TN, U, TPI> _buffer;
+    private protected RMTensor4Buffer4x4<TT, TN, U, TPI> _buffer;
 
-    public RMTensorField3() { }
+    public TensorFieldR4x4() { }
 
-    public Func<TT, AutoDiffTensor3<TN, U, TPI>, Variable<TN, U>> this[int i]
+    public Func<TT, AutoDiffTensor4<TN, U, TPI>, Variable<TN, U>> this[int i, int j]
     {
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        get => _buffer[i];
+        get => _buffer[i][j];
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        set => _buffer[i] = value;
+        set => _buffer[i][j] = value;
     }
 
-    /// <inheritdoc cref="RMTensorField2{TT, TN, U, TIP, TPI}.Compute{TIN}(TT, AutoDiffTensor2{TN, U, TPI})"/>
-    public Tensor<Vector3<TN, U>, TN, U, Index<TIP, TIN>> Compute<TIN>(TT tape, AutoDiffTensor3<TN, U, TPI> point)
-        where TIN : IIndexName
+    /// <inheritdoc cref="TensorFieldR2x2{TT, TN, U, TI1P, TI2P, TPI}.Compute{TI1N, TI2N}(TT, AutoDiffTensor2{TN, U, TPI})"/>
+    public Tensor<Matrix4x4<TN, U>, TN, U, Index<TI1P, TI1N>, Index<TI2P, TI2N>> Compute<TI1N, TI2N>(TT tape, AutoDiffTensor4<TN, U, TPI> point)
+        where TI1N : IIndexName
+        where TI2N : IIndexName
     {
-        Vector3<TN, U> result = new();
-        for (int i = 0; i < 3; i++)
+        tape.IsTracking = false;
+
+        Matrix4x4<TN, U> result = new();
+        for (int i = 0; i < 4; i++)
         {
-            if (_buffer[i] is Func<TT, AutoDiffTensor3<TN, U, TPI>, Variable<TN, U>> function)
-                result[i] = function(tape, point).Value;
+            for (int j = 0; j < 4; j++)
+            {
+                if (_buffer[i][j] is Func<TT, AutoDiffTensor4<TN, U, TPI>, Variable<TN, U>> function)
+                    result[i, j] = function(tape, point).Value;
+            }
         }
-        return new Tensor<Vector3<TN, U>, TN, U, Index<TIP, TIN>>(result);
+
+        tape.IsTracking = true;
+        return new Tensor<Matrix4x4<TN, U>, TN, U, Index<TI1P, TI1N>, Index<TI2P, TI2N>>(result);
     }
 }
 
 internal static partial class Buffers
 {
-    [InlineArray(3)]
-    internal struct RMTensor3Buffer3<TT, TN, U, TPI>
+    [InlineArray(4)]
+    internal struct RMTensor4Buffer4x4<TT, TN, U, TPI>
         where TT : ITape<TN, U>
         where TN : IComplex<TN, U, U>, IDifferentiableFunctions<TN>
         where U : IBinaryFloatingPointIeee754<U>, IMinMaxValue<U>
         where TPI : IIndex
     {
-        private Func<TT, AutoDiffTensor3<TN, U, TPI>, Variable<TN, U>> _element;
+        private RMTensor4Buffer4<TT, TN, U, TPI> _element;
     }
 }
